fix: reordering of tests
