+++ tests/conftest.py
@@ -2,6 +2,8 @@ from pathlib import Path
 
 import pytest
 
+from pyc2ray.load_extensions import libasora
+
 
 @pytest.fixture(scope="session")
 def test_dir() -> Path:
@@ -13,3 +15,12 @@ def test_dir() -> Path:
 def data_dir(test_dir: Path) -> Path:
     """Return the path to the data folder for tests"""
     return test_dir / "data"
+
+
+@pytest.fixture
+def init_device():
+    if libasora is not None:
+        libasora.device_init()
+    yield
+    if libasora is not None:
+        libasora.device_close()
